requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0

numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2

markdown==3.6
markdown2

python-dotenv==1.0.1






PyYAML

weasyprint

torch
torchview
torchsummary

google-api-python-client==2.126.0
google-auth==2.28.2
google-auth-oauthlib==1.2.0

uvicorn==0.29.0
gunicorn==21.2.0

# optional if you use SharedMemory
multiprocess==0.70.16

langchain
langchain-core
langchain-chroma
langchain-google-genai
langgraph

chromadb==0.4.24